Add instructions for changing applicationId
The applicationId is used for the package name of the apk.  Changing
the package name in AndroidManifest.xml does not change the package
name of the apk, it is only for resources and code, so the theme
designer needs to be sure to change the applicationId in order to
set their package name.

Change-Id: Ifa511950d904f7b646c2e1aca9aca8de213e3151